+++ src/app/templates/server/server.js
@@ -6,7 +6,8 @@ const apiRouter = express.Router();
 
 apiRouter.all('/*', (req, res, next) => {
  res.header("Access-Control-Allow-Origin", "*");
- res.header("Access-Control-Allow-Headers", "X-Requested-With");
+ res.header("Access-Control-Allow-Headers", "Content-Type, X-Requested-With");
+ res.header("Access-Control-Allow-Methods", "GET,POST,PUT,DELETE,OPTIONS");
 
  next();
  });
@@ -23,3 +24,10 @@ const server = app.listen(port, () => {
 
     console.log('Example app listening at http://%s:%s', host, port)
 });
+
+function randomString(length) {
+    const chars = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+    var result  = '';
+    for (var i = length; i > 0; --i) result += chars[Math.floor(Math.random() * chars.length)];
+    return result;
+}