+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## HEAD
 
+## v0.1.4
+###### _Jan 21, 2016_
+
+##### General
+- New entity and api handlers generated on :router subgenerator
+- New actions for add/remove entity and reducers for them
+- Sample data table generated with new route
+
+
 ## v0.1.3
 ###### _Jan 14, 2016_